fix error on counter data type in powf wgsl file, +sqrt, sqrt old code has been remove

+++ src/arr_o_gpu/compute_shaders/function/powf/powf_contiguous.wgsl
@@ -18,7 +18,7 @@ var<storage, read_write> heap: array<f32>;
 var<uniform> execute_args: array<ArrayMetadata32, 3>;
 // // static_cache
 struct CounterPowf{
-    counter: f32,
+    counter: u32,
     scalar: f32,
     index: u32,
     padding: u32,
